pyproject.toml
[tool.poetry]
name = "iven"
version = "0.1.0"
description = "A simple bot powered by ariadne"
authors = ["yinlinghunjiang <1581647952@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
graia-ariadne = {extras = ["standard"], version = "^0.9.6"}
graia-saya = "^0.0.16"
urllib3 = "^1.26.12"
ariadne = "^0.17.0"
asyncio = "^3.4.3"
opencv-python = "^4.6.0.66"
aiohttp = "^3.8.3"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna-tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
